Fix bundle size regression
